pyproject.toml
[project]
name = "fast_zero"
version = "0.1.0"
description = "FastAPI do zero!"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "aiosqlite>=0.22.1",
  "alembic>=1.17.2",
  "fastapi[standard]>=0.127.1",
  "logfire[fastapi]>=4.16.0",
  "pwdlib[argon2]>=0.3.0",
  "pydantic-settings>=2.12.0",
  "pydantic[email]>=2.12.5",
  "pyjwt>=2.10.1",
  "sqlalchemy[asyncio]>=2.0.45",
  "tzdata>=2025.3",
]

[dependency-groups]
dev = [
  "pre-commit>=4.5.1",
  "pyrefly>=0.46.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.14.10",
  "taskipy>=1.14.1",
]

[tool.coverage.run]
core = "ctrace"
concurrency = ["thread", "greenlet"]

[tool.logfire]
pydantic_plugin_record = "all"

[tool.pyrefly]
project-includes = ["app/**/*.py"]
project-excludes = ["migrations/**", "tests/**"]
search-path = ["app"]
python-version = "3.12"

[tool.pyrefly.errors]
not-async = false

[tool.ruff]
line-length = 79
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["FAST", "I", "F", "E", "W", "PL", "PT"]

[tool.ruff.lint.per-file-ignores]
'__init__.py' = ["F401", "F403"]
'test_*.py' = ["S101", "ARG", "S106", "S105"]
'conftest.py' = ["ARG", "S106", "S105"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = [
  "pydantic.BaseModel",
  "sqlalchemy.orm.DeclarativeBase",
  "pydantic_settings.BaseSettings",
]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.taskipy.tasks]
format = "ruff format ."
lint = "ruff check ."
post_lint = "ruff check . --diff"
post_test = "coverage html"
pre_format = "ruff check . --fix"
pre_lint = "pyrefly check 'app'"
pre_test = "task format"
run = "fastapi dev app/main.py"
test = "pytest -s -x --cov=app -vv"
